sync_mysql2elasticsearch/requirements.txt
elasticsearch==7.14.0
mysql-replication==0.21
PyMySQL==0.9.3